+++ CHANGELOG.md
@@ -6,7 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.9.3] - 2025-07-21
+- feat: `ZIO Test` on `Scala.js` and `Scala Native`;
+- chore: dependency updates;
+- cleanup;
+
 ## [0.9.2] - 2025-07-18
+- feat: `specs2` on `ScalaNative`;
 - chore: Gradle 9.0.0-rc-3;
 - chore: dependency updates;
 - cleanup;